+++ packages/pluggableWidgets/web-view-native/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Structure mode preview
+
 ## [3.1.1] - 2022-2-21
 
 ### Fixed